Removed classes/scraper/saucenao.py   - copy/pasted SauceNao class under classes/saucenao.py for reference
